% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WhatsappChatCleaner.R
\name{clean_whatsapp_chat}
\alias{clean_whatsapp_chat}
\title{Cleans whatsapp chat data}
\usage{
clean_whatsapp_chat(filepath)
}
\arguments{
\item{filepath}{, string, enter file path or file name if present in the same directory}
}
\description{
This function takes the raw chat file and makes it clean for analysis.
}
\examples{
clean_data <- clean_whatsapp_chat("whatsap_friends.txt")
}
\keyword{media}

