% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WhatsappChatCleaner.R
\name{emoji_loader}
\alias{emoji_loader}
\title{Load Emojis from file}
\usage{
emoji_loader(filepath)
}
\arguments{
\item{filepath}{, string, enter file path or file name if present in the same directory}
}
\description{
This function loads the emojis from csv file and cleans them for usage.
}
\examples{
emoji_all = emoji_loader("whatsapp_emoji_all.csv")
}

