% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WhatsappChatCleaner.R
\name{media_remove}
\alias{media_remove}
\title{Remove media records}
\usage{
media_remove(clean_data)
}
\arguments{
\item{clean_data}{, data frame}
}
\description{
This function removes all the media messages rows from the data frame.
}
\examples{
clean_data_without_media <- media_remove(clean_data)
}
\keyword{media}

